fix(tiling): only a settled read may confirm from the baseline
An echo reading equal to the pre-ask baseline is ambiguous: "the
app refused" and "the app has not redrawn yet" are the same
frame. Under load the second is ordinary for any app — the fast
ones included — so the baseline arm recorded our own latency as
the app's limit.

Device capture 2026-08-28 (load average 9.7, macOS 26.6.2): one
false bound per press, each at exactly the pre-press width (984,
954, 924, 894 — one resize step apart), heights all 1011, which
is the slot geometry rather than any app's limit. Sixteen
confirmations against a single candidate in eight minutes. The
false bound then fed ScrollSlotDomain, which refused the next
press and cued a size-limit pill naming a limit the window was
nowhere near; a window resized both ways ended pinned between a
false minimum and a false maximum.

The arm exists so a size held unchanged through a whole
animation completes the ladder in one observation rather than
two. That intent is kept — the settle probe still confirms it in
a single read — but only the channel that waited out the grace
may take it. A genuinely refusing window is now learned one
probe grace after its animation settles.

fixes #1083
